+++ sdk-lib/src/androidTest/java/cash/z/ecc/android/sdk/integration/service/ChangeServiceTest.kt
@@ -12,8 +12,9 @@ import cash.z.ecc.android.sdk.internal.service.LightWalletGrpcService
 import cash.z.ecc.android.sdk.internal.service.LightWalletService
 import cash.z.ecc.android.sdk.internal.twig
 import cash.z.ecc.android.sdk.model.BlockHeight
+import cash.z.ecc.android.sdk.model.LightwalletdServer
+import cash.z.ecc.android.sdk.model.ZcashNetwork
 import cash.z.ecc.android.sdk.test.ScopedTest
-import cash.z.ecc.android.sdk.type.ZcashNetwork
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
@@ -35,13 +36,14 @@ import org.mockito.Spy
 class ChangeServiceTest : ScopedTest() {
 
     val network = ZcashNetwork.Mainnet
+    val lightwalletdServer = LightwalletdServer("mainnet.lightwalletd.com", 9087)
 
     @Mock
     lateinit var mockBlockStore: CompactBlockStore
     var mockCloseable: AutoCloseable? = null
 
     @Spy
-    val service = LightWalletGrpcService(context, network)
+    val service = LightWalletGrpcService(context, network, lightwalletdServer)
 
     lateinit var downloader: CompactBlockDownloader
     lateinit var otherService: LightWalletService
@@ -50,7 +52,7 @@ class ChangeServiceTest : ScopedTest() {
     fun setup() {
         initMocks()
         downloader = CompactBlockDownloader(service, mockBlockStore)
-        otherService = LightWalletGrpcService(context, "lightwalletd.electriccoin.co")
+        otherService = LightWalletGrpcService(context, LightwalletdServer("lightwalletd.electriccoin.co", 9087))
     }
 
     @After