+++ sdk-lib/src/test/java/cash/z/ecc/android/sdk/model/LightwalletdServerTest.kt
@@ -1,0 +1,16 @@
+package cash.z.ecc.android.sdk.model
+
+import org.junit.Test
+import kotlin.test.assertTrue
+
+class LightwalletdServerTest {
+    @Test
+    fun requireSecureMainnet() {
+        assertTrue(LightwalletdServer.Mainnet.isSecure)
+    }
+
+    @Test
+    fun requireSecureTestnet() {
+        assertTrue(LightwalletdServer.Testnet.isSecure)
+    }
+}